Bump and assemble refactoring
This commit bump software to latest version. Main change is
Python 3.9. Minor part of this a lot of different small packages.

This commit adds rust and cargo as build time dependencies bescause
of new parts in crypto engine (PEP517).

This commit also change logic of assembling. Everything related to
build is in assemble dir as it was done for adss_core previously.